fusion-retrieval/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# LlamaIndex dependencies
llama-index==0.14.8
llama-index-vector-stores-faiss==0.5.1
llama-index-retrievers-bm25==0.6.5
llama-index-llms-openai==0.6.9
llama-index-embeddings-openai==0.5.1
llama-index-readers-file==0.5.5

# Vector store
faiss-cpu==1.13.0

# Configuration
python-dotenv==1.2.1
pyyaml==6.0.1

# Utilities
numpy==1.26.2


